First iteration of a simple replacement for WebMatrix.WebSecurity.
Maybe some day it'll become worth it to pay for
sql storage on Azure for this and I can just use
the default.

I'm assuming I will have to come back to this
from time to time. Not happy with the solution
of just switching namespace rather than using a
new class name but it was easy.

Also commented out most of the OAuth related
code, I'll get back to it when the time comes.